requirements.txt
aiogram==3.21.0
pyrogram==2.0.106
aiosqlite==0.21.0
python-dotenv==1.1.1
aiofiles==24.1.0
aiohttp==3.12.14
certifi==2025.7.14
magic-filter==1.0.12
pydantic==2.11.7
typing-extensions==4.14.1
pyaes==1.6.1
pysocks==1.7.1 